requirements.txt
numpy==1.25.2
pexpect==4.9.0
ptyprocess==0.7.0
python-dateutil==2.8.2
pytz==2023.3
pywebio==1.8.2
PyYAML==6.0.1
six==1.16.0
tornado==6.3.3
tzdata==2023.3
ua-parser==0.18.0
user-agents==2.2.0
